follow single responsibility principle
