+++ Slim/Handlers/NotAllowed.php
@@ -36,22 +36,22 @@ class NotAllowed extends AbstractHandler
         if ($request->getMethod() === 'OPTIONS') {
             $status = 200;
             $contentType = 'text/plain';
-            $output = $this->renderPlainOptionsMessage($methods);
+            $output = Render::make('PlainOptionsMessage', $methods);
         } else {
             $status = 405;
             $contentType = $this->determineContentType($request);
             switch ($contentType) {
                 case 'application/json':
-                    $output = $this->renderJsonNotAllowedMessage($methods);
+                    $output = Render::make('JsonNotAllowedMessage', $methods);
                     break;
 
                 case 'text/xml':
                 case 'application/xml':
-                    $output = $this->renderXmlNotAllowedMessage($methods);
+                    $output = Render::make('XmlNotAllowedMessage', $methods);
                     break;
 
                 case 'text/html':
-                    $output = $this->renderHtmlNotAllowedMessage($methods);
+                    $output = Render::make('HtmlNotAllowedMessage', $methods);
                     break;
                 default:
                     throw new UnexpectedValueException('Cannot render unknown content type ' . $contentType);
@@ -67,81 +67,5 @@ class NotAllowed extends AbstractHandler
                 ->withHeader('Content-type', $contentType)
                 ->withHeader('Allow', $allow)
                 ->withBody($body);
-    }
-
-    /**
-     * Render PLAIN message for OPTIONS response
-     *
-     * @param  array                  $methods
-     * @return string
-     */
-    protected function renderPlainOptionsMessage($methods)
-    {
-        $allow = implode(', ', $methods);
-
-        return 'Allowed methods: ' . $allow;
-    }
-
-    /**
-     * Render JSON not allowed message
-     *
-     * @param  array                  $methods
-     * @return string
-     */
-    protected function renderJsonNotAllowedMessage($methods)
-    {
-        $allow = implode(', ', $methods);
-
-        return '{"message":"Method not allowed. Must be one of: ' . $allow . '"}';
-    }
-
-    /**
-     * Render XML not allowed message
-     *
-     * @param  array                  $methods
-     * @return string
-     */
-    protected function renderXmlNotAllowedMessage($methods)
-    {
-        $allow = implode(', ', $methods);
-
-        return "<root><message>Method not allowed. Must be one of: $allow</message></root>";
-    }
-
-    /**
-     * Render HTML not allowed message
-     *
-     * @param  array                  $methods
-     * @return string
-     */
-    protected function renderHtmlNotAllowedMessage($methods)
-    {
-        $allow = implode(', ', $methods);
-        $output = <<<END
-<html>
-    <head>
-        <title>Method not allowed</title>
-        <style>
-            body{
-                margin:0;
-                padding:30px;
-                font:12px/1.5 Helvetica,Arial,Verdana,sans-serif;
-            }
-            h1{
-                margin:0;
-                font-size:48px;
-                font-weight:normal;
-                line-height:48px;
-            }
-        </style>
-    </head>
-    <body>
-        <h1>Method not allowed</h1>
-        <p>Method not allowed. Must be one of: <strong>$allow</strong></p>
-    </body>
-</html>
-END;
-
-        return $output;
     }
 }
