+++ Slim/Handlers/Renderables/HtmlNotAllowedMessage.php
@@ -1,0 +1,51 @@
+<?php
+namespace Slim\Handlers\Renderables;
+
+use Slim\Interfaces\RenderableInterface;
+
+class HtmlNotAllowedMessage implements RenderableInterface
+{
+
+    public function render($args)
+    {
+        return $this->renderHtmlNotAllowedMessage($args[0]);
+    }
+
+
+    /**
+     * Render HTML not allowed message
+     *
+     * @param  array                  $methods
+     * @return string
+     */
+    protected function renderHtmlNotAllowedMessage($methods)
+    {
+        $allow = implode(', ', $methods);
+        $output = <<<END
+<html>
+    <head>
+        <title>Method not allowed</title>
+        <style>
+            body{
+                margin:0;
+                padding:30px;
+                font:12px/1.5 Helvetica,Arial,Verdana,sans-serif;
+            }
+            h1{
+                margin:0;
+                font-size:48px;
+                font-weight:normal;
+                line-height:48px;
+            }
+        </style>
+    </head>
+    <body>
+        <h1>Method not allowed</h1>
+        <p>Method not allowed. Must be one of: <strong>$allow</strong></p>
+    </body>
+</html>
+END;
+
+        return $output;
+    }
+}
