+++ Slim/Handlers/Renderables/JsonErrorMessage.php
@@ -1,0 +1,50 @@
+<?php
+namespace Slim\Handlers\Renderables;
+
+use Slim\Interfaces\RenderableInterface;
+
+class JsonErrorMessage implements RenderableInterface
+{
+
+    private $displayErrorDetails;
+
+    public function render($args)
+    {
+        $this->displayErrorDetails = $args[1];
+        return $this->renderJsonErrorMessage($args[0]);
+    }
+
+
+    /**
+     * Render JSON error
+     *
+     * @param \Exception $exception
+     *
+     * @return string
+     */
+    protected function renderJsonErrorMessage(\Exception $exception)
+    {
+        $error = [
+            'message' => 'Slim Application Error',
+        ];
+
+        if ($this->displayErrorDetails) {
+            $error['exception'] = [];
+
+            do {
+                $error['exception'][] = [
+                    'type' => get_class($exception),
+                    'code' => $exception->getCode(),
+                    'message' => $exception->getMessage(),
+                    'file' => $exception->getFile(),
+                    'line' => $exception->getLine(),
+                    'trace' => explode("\n", $exception->getTraceAsString()),
+                ];
+            } while ($exception = $exception->getPrevious());
+        }
+
+        return json_encode($error, JSON_PRETTY_PRINT);
+    }
+
+
+}
