+++ Slim/Handlers/Renderables/XmlNotFoundOutput.php
@@ -1,0 +1,14 @@
+<?php
+namespace Slim\Handlers\Renderables;
+
+use Slim\Interfaces\RenderableInterface;
+
+
+class XmlNotFoundOutput implements RenderableInterface
+{
+
+    public function render($args)
+    {
+        return '<root><message>Not found</message></root>';
+    }
+}
